requirements.txt
django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
faker==17.6.0
django-rest-swagger==2.2.0
